0004026: Registration server gets "Removing identity because registration is required"

+++ symmetric-core/src/main/java/org/jumpmind/symmetric/service/impl/PushService.java
@@ -221,7 +221,7 @@ public class PushService extends AbstractOfflineDetectorService implements IPush
         } catch (Exception ex) {
             processInfo.setStatus(ProcessStatus.ERROR);
             fireOffline(ex, remote, status);
-            if (isRegistrationRequired(ex)) {
+            if (isRegistrationRequired(ex) && !parameterService.isRegistrationServer()) {
                 log.info("Removing identity because registration is required");
                 nodeService.deleteIdentity();
                 nodeService.findIdentity(false);
